Adding Students Feature: check if email is taken and handle delete by nonexistent id

+++ src/main/java/com/joaco/fullstackcourse/student/StudentService.java
@@ -1,7 +1,10 @@
 package com.joaco.fullstackcourse.student;
 
+import com.joaco.fullstackcourse.student.exception.BadRequestException;
+import com.joaco.fullstackcourse.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+
 import java.util.*;
 
 @AllArgsConstructor
@@ -14,16 +17,19 @@ public class StudentService {
         return studentRepository.findAll();
     }
 
-
     public void addStudent(Student student) {
-        //TODO check if email is taken
-
+        Boolean existsEmail = studentRepository.selectExistsEmail(student.getEmail());
+        if (existsEmail) {
+            throw new BadRequestException("Email " + student.getEmail() + " taken");
+        }
         studentRepository.save(student);
 
     }
 
     public void deleteStudent(Long studentId) {
-        //TODO check if student exist
+        if (!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException("Student with id " + studentId + " does not exists");
+        }
         studentRepository.deleteById(studentId);
     }
 }
